+++ CHANGELOG.md
@@ -11,6 +11,7 @@ v1.10.0 (unreleased)
     blocks requests until a peer becomes available.
 -   Add support for reading ShardKey, RoutingKey, and RoutingDelegate for
     inbound http and tchannel calls.
+-   Move encoding/x/protobuf to encoding/protobuf.
 
 
 v1.9.0 (2017-06-08)